backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings
pyyaml
docxtpl
python-docx
pytesseract
pillow
pillow-heif
pypdf
openpyxl
numpy
opencv-python-headless
paddlepaddle
paddleocr
rapidocr-onnxruntime
